mod_archive: Fix duplicate messages stored
Message exchanged between users on the same server would be stored
twice. This is because both the message hook and the pre-message hook
would store for both users.

The solution is to make sure the pre-message hook only store for the
'from' user, and the post message hook only store for the 'to' user.